refactored routes in ui_server.py for easier testing (and overview)
