+++ scanomatic/ui_server/status_ui.py
@@ -1,0 +1,35 @@
+from flask import send_from_directory, redirect, jsonify
+
+from scanomatic.io.paths import Paths
+from .general import convert_path_to_url
+
+
+def add_routes(app, rpc_client):
+
+    @app.route("/status")
+    @app.route("/status/<status_type>")
+    def _status(status_type=""):
+
+        if status_type != "" and not rpc_client.online:
+            return jsonify(success=False, reason="Server offline")
+
+        if status_type == 'queue':
+            return jsonify(success=True, data=rpc_client.get_queue_status())
+        elif 'scanner' in status_type:
+            return jsonify(success=True, data=rpc_client.get_scanner_status())
+        elif 'job' in status_type:
+            data = rpc_client.get_job_status()
+            for item in data:
+                if item['type'] == "Feature Extraction Job":
+                    item['label'] = convert_path_to_url("", item['label'])
+                if 'log_file' in item and item['log_file']:
+                    item['log_file'] = convert_path_to_url(
+                        "/logs/project", item['log_file'])
+            return jsonify(success=True, data=data)
+        elif status_type == 'server':
+            return jsonify(success=True, data=rpc_client.get_status())
+        elif status_type == "":
+
+            return send_from_directory(Paths().ui_root, Paths().ui_status_file)
+        else:
+            return jsonify(succes=False, reason='Unknown status request')
